+++ src/ray.zig
@@ -4,13 +4,13 @@ const Vec3 = color.Vec3;
 pub const Ray = struct {
     orig: Vec3,
     dir: Vec3,
-    tm: f32 = 0,
+    tm: f64 = 0,
 
     pub fn init(origin_: Vec3, direction_: Vec3) Ray {
         return .{ .orig = origin_, .dir = direction_, .tm = 0 };
     }
 
-    pub fn initTimed(origin_: Vec3, direction_: Vec3, time_: f32) Ray {
+    pub fn initTimed(origin_: Vec3, direction_: Vec3, time_: f64) Ray {
         return .{ .orig = origin_, .dir = direction_, .tm = time_ };
     }
 
@@ -22,11 +22,11 @@ pub const Ray = struct {
         return self.dir;
     }
 
-    pub fn time(self: Ray) f32 {
+    pub fn time(self: Ray) f64 {
         return self.tm;
     }
 
-    pub fn at(self: Ray, t: f32) Vec3 {
+    pub fn at(self: Ray, t: f64) Vec3 {
         return self.orig.add(self.dir.scale(t));
     }
 };
